Add new method `sampleValueAndDerivatives` to `BicubicSplineInterpolation` to sample the value and derivatives together, reducing the number of spline evaluations from three to two. 

I also removed the temporary vectors from each method, replacing them with member vectors. This alone sped up each interpolation by between 5% and 10% (depending on number of points in vectors).

Some timing results for 1 million samples for the value and both first derivatives using this new method to calculate all at once, compared to the old method of evaluating each separately:

|              | Before | After | Speedup |
|--------|--------|------|----------|
| BicubicSplineInterpolation | 9.28251s | 5.77597s | 1.6 |
| TabulatedFluidProperties | 15.1424s | 9.29558s | 1.6 |

These results depend on the number of grid points etc, but I typically saw this degree of reduction in time +/- a few percent, which makes sense as it is basically reducing the calculation by a third.

Close #8936
Closes #10995